Move transcoding out of ImplicitTileset
